pr_gettext: stop tracking (and printing) untranslated strings
Do not collect anymore the information of untranslated strings, since it
is only used to print all of them at the end of the message extraction.

While I understand the reason why this was done (i.e. to discover
potentially untranslated user strings), this is something that just
causes lots of output in complex projects, since strings are basically
used for many non-UI tasks (say Str, Sys.command, etc), so the actual
result is that there are lots of false positive.  Also, this is not
something xgettext (from GNU gettext) does.